backend/deploy/src/requirements.txt
flask
gunicorn
flask_cors
flask_sqlalchemy
lxml
psycopg2-binary
sqlalchemy
pyyaml
sqlalchemy_utils
sqlalchemy_serializer
xmltodict
gigachat
# Статистика
numpy
scipy